sym-prop: Fix matching with deep expression patterns (#4483)
Let's say you want to match pattern:

    return <... name.contains("...") ...>;

against this code:

    boolean b1 = !name.contains("../");
    boolean b2 = !name.contains("..\\");
    return b1 && b2;

It is not enough to compute the subexpressions of `b1 && b2` and then try
matching them with `name.contains("...")`. This does not work. We need to
consider symbolic values *while* computing the subexpressions, so that
`name.contains("../")` and `name.contains("..\\")` are enumerated.

Also allow special ids to be symbolically propagated.

Closes #3207

test plan:
make test # test included

changelog:
Symbolic propagation didn't exist in 0.76.0, but may have ended up there
by mistake when merging the release-0.76.0 branch. The feature never made
it into a release announcement, so I'm adding it for the next one.

Co-authored-by: Yoann Padioleau <pad@returntocorp.com>