Only report parse errors once per file (#4486)
* Only report parse errors once per file

Currently, when semgrep runs multiple rules on the same file with a parse error, it will warn on each rule/file combination. This PR fixes that.

* Revert "Only report parse errors once per file"

This reverts commit c867f2e64082d027901cfa7daf95f714a0aa625c.

* Only report parse errors once per file

Currently, when semgrep runs multiple rules on the same file with a parse error, it will warn on each rule/file combination. This PR fixes that.

* Add test for deduplication

* Rename targets folder for test

* Changelog

* Update snapshots

* Sort rules for stable testing

* Changelog

* rule_id now unnecessary