Send all rules to semgrep-core
Instead of sending the rules one at a time to semgrep-core, send them all, thus only invoking semgrep-core once. We expect this to cause a significant speed up for diff scans (or other scans that run many rules on only a few bytes of target).

This also requires changing the return object. Matches and errors are still returned in a flat list as before, with the rule id annotated. The profiling information had to be grouped differently. A few fields are lost (run time is now known for each file, not each file and each rule). Also, the progress bar is lost.

TODOs:
- This PR makes minimal changes necessary to send all the rules. In the future, we will probably want to refactor the code to operate file-wise (which is how results are outputted).
- This PR does not solve the error reporting problem for rules. Rules are still parsed twice, and the rules sent to semgrep-core are reindented. To change that, we will need to change pattern-ref and also allow semgrep-core to take multiple rules files (or else perform a transformation on the Python side)