Transfer Graph_code_AST from codegraph to semgrep
This poor's man name resolver will be useful for
semgrep variants that want to perform some global analysis

This code used to be in codegraph (one of pad's personal github
project), but we need it now in semgrep and we don't want to
add dependencies towards codegraph. In the long term
we could move out lots of code out of semgrep in a
big AST-generic-parsers-and-utilities separate package.

test plan:
make