core: pcre: Prevent Pcre engine from 'hanging" Semgrep (#4648)
The PCRE engine does not allow setting timeouts and the existing PCRE
settings set very generous limits. Given the right regex and file, the
PCRE matching can take a long time to run, and `Common.set_timeout`
won't work as PCRE is calling a C library! As a result, Semgrep appeared
to hang.

Here we lower the match-attempts and recurson limits for the PCRE engine,
hoping to avoid such "hangs". For now, when the limits are reached, we
return as if the regex did not match, but we should probably report these
situations as timeout.

test plan:
$ semgrep-core -lang regex -rules perf/input/semgrep_targets.yaml perf/input/semgrep_targets.txt
 #^ now returns almost instantly whereas previously it would take a long time
$ cd perf && ./run-perf-suite # test included