--validate checks metavariables used by metavariable-x
When metavariable-pattern or metavariable-regex is used on a metavariable that is not valid (not in any pattern-and clause or all pattern-or clauses), emit an error.

Also, remove the redundant "duplicate pattern" and "unsatisfiable" checks that are already covered by semgrep-rule-lints. We previously could not because otherwise there would have been no test that the OCaml checks are being used.

Test plan: `semgrep --validate --config tests/OTHER/metachecks/unused-metavariable.yaml`

TODO:
[ ] fix --validate message for semgrep-cli
[ ] add --validate tests to semgrep-cli
[ ] improve message for the unused metavariable check to be more helpful