[HCL] Fix CST-to-Generic translation of heredocs (#4660)
* [HCL] Fix CST-to-Generic translation of heredocs

Heredocs are parsed as a list of tokens without any newline or
whitespace. But heredocs may represent code that we want to analyze with
metavariable-regex or metavariable-pattern, so we must parse these
strings correctly. Ideally, we should preserve these newlines and
whitespace already during parsing... but for now we just "restore" them
using token location info.

Closes #4582

test plan:
$ semgrep-core -lang hcl -dump_ast tests/OTHER/rules/terraform_nested_yaml.tf
 #^ see how the definition of content is represented in Generic
$ make test # test included

* Emma's review