fix(autofix): Stop crashing when dry-running two fixes for the same file
When there are two autofixes for the same file,
and the first one shifts entire lines around,
Semgrep is smart and takes that shift into account for the second file.
But! If we're dry-running the fixes, we'd still try to compensate
for the changes that are not actually written to disk.
The second autofix would then crash.

The fix here is to get rid of the shift-adjustments when dry-running.

Please also enjoy our complimentary pack of eleven brand-new test cases.