fix(core_runner): performance regression with --debug
When semgrep was run with --debug we were seeing a significant performance
regression than without. The root cause was in core_runner we were appending
stdout and stderr to a string which was running in O(n^3) for n line length.

With this PR we append to a list O(1) and do a single join at the end
removing the perf regression