Refactor TargetManager
The primary goal of this refactor is to split off a Target class that
maps to the marked user input strings:

semgrep -f auto foo.py bar/
                ^^^^^^ ^^^^
Target knows how to return all paths that such a string refers to,
according to one of three strategies (os.walk, git ls-files, or git diff).

The refactoring surfaced inconsistencies in our targeting behavior,
which were also addressed.

The implementation of the git diff strategy,
and the CHANGELOG entries about the changes from this refactoring
are in the next commit.