taint-mode: Stop unifying source-sink metavariables (#4901)
* taint-mode: Stop unifying source-sink metavariables

At least it will not be the default. Forcing unification of source-sink
metavariables turned out to be confusing even for r2c's security-research
team!

We never had a real use case for unification (although it seems there are
some), and we just wanted to expose the metavariables to the rule message,
which we can achieve in other ways. Here we chose to do the union of both
metavariable sets, but prefer the sink ones in case of collision.

And restore most of "taint-mode: Propagate metavariables correctly (#4877)"
We still have to propagate metavariables correctly!

Closes #4464 (again)

test plan:
make test # test included

* Emma's review