Rules with invalid patterns fail fast (#4969)
* Rules with invalid patterns fail fast

In 0.86.5, semgrep-core started collecting the invalid rules to send at the end. This is nice if we want to fail open, but we haven't implemented that in semgrep yet. For now, just throw the exception but leave the infrastructure to fail open.

Test plan:

rule_file.yaml
```
rules:
- id: println
  pattern: System.out.println(...);
  message: Using println
  languages: [java]
  severity: WARNING
- id: console_log
  pattern: console.l..);
  message: Using console.log
  languages: [js]
  severity: WARNING
```

Run `semgrep --config rule_file.yaml .`, expect a crash

* Fixup

* Tests and changelog