feat(swift): Add Semgrep ellipsis to Swift patterns (#5193)
Swift has an existing unbounded range operator that is a valid
expression. When we are parsing a pattern, we should interpret this as a
Semgrep ellipsis.

This currently needs the parentheses in the test pattern because without
that, it's not a valid parse for the unbounded range operator. After
https://github.com/returntocorp/ocaml-tree-sitter-semgrep/pull/302 makes
its way here, we'll be able to omit the parentheses.

Test plan:

Automated tests