Add support for defining "metatypes" to use in rules
This is an experiment based on a conversation with security. We allow
users to pass in a file (like equivalences, but specifically for types)
to define a metatype. When they use this metatype in a rule, we also
check whether any of the metatype's subtypes are present in the code.
In addition, we still check for the metatype, to allow the name to be
reused.

Test plan:

Define three files:

```
➜  metatypes cat log_rule.yaml
rules:
- id: secret-log
  pattern: (LoggerType $LOGGER).log(secret);
  message: Don't log secrets!
  languages:
  - java
  severity: ERROR
➜  metatypes cat app.java
public class Test {

    public static void main(String[] args) {
        Logger log1 = new Logger();
        Log4j.Logger log2 = new Log4j.Logger();
        MyLogger log3 = new MyLogger();
        SafeLogger log4 = new SafeLogger();

        String secret = "Keep this secret";

        // MATCH:
        log1.log(secret);
        log1.log("innocuous");

        // TODO
        log2.log(secret);
        log2.log("innocuous");

        // MATCH:
        log3.log(secret);
        log3.log("innocuous");

        // Ok:
        log4.log(secret);
        log4.log("innocuous");
    }

}
➜  metatypes cat metatypes.yaml
types:
- LoggerType:
    - Logger
    - Log4j.Logger
    - MyLogger
```

```
➜  metatypes sc -lang java -rules log_rule.yaml app.java
```

```
➜  metatypes sc -lang java -rules log_rule.yaml app.java -metatypes metatypes.yaml
app.java:11 with rule secret-log
         log1.log(secret);
app.java:17 with rule secret-log
         log3.log(secret);
```