fix: Attempt a match between Id and IdQualified
It's possible for DeepSemgrep to match in this case when the pattern is
an `Id` and the source is an `IdQualified`, but a parent of the source
name ultimately resolves to a matching `Id`.

At some point we'll likely have to add another case so that
`IdQualified` can match `Id`.

Test plan: Automated tests