Print parse errors in dump_ast
In addition to printing the warning that the file failed to fully parse,
print the parse errors that actually led to that warning. This is useful
when investigating parse errors.

Test plan:

`$ semgrep-core -tree_sitter_only -lang php -dump_ast filewitherrors.php`

See the parse errors reported on the CLI