fix(ruby): Use QDots in IDQualified when possible
When resolved names go through the matching infrastructure in
`Generic_vs_generic.m_name`, they are transformed into AST nodes via the
`name_of_ids` helper function. This transformation is lossy, and in
particular, it turnes a series of multiple identifiers into an
`IdQualified` node using `QDots` as the qualifier.

Unfortunately, in Ruby, `QDots` never appears in the AST. Instead, even
a plain qualified name is represented using `QExpr`. This means that the
pattern `Foo::Bar` is represented with `QExpr`, and can be compared
against AST constructed from a resolved name using `QDots`. This causes
a match failure in every case.

We could add a case to `m_qualifier` to extract names from the
special-case of `QExpr` that contains only a name, but I've already had
to do this in one other place in DeepSemgrep and I suspect that
continuing down that path would lead to numerous places where we would
need to handle this special case.

Instead, let's handle the special case once, when we construct the
generic AST. For the common case, where the qualifier is in fact a
series of identifiers, this will make downstream code behave similarly
for Ruby as it does for other languages that use `QDots` exclusively. In
particular, this allows DeepSemgrep to match a pattern `Foo::Bar`
against something that is not exactly `Foo::Bar` but resolves to it.

Test plan:

Automated tests

`semgrep-core -lang ruby -dump_ast test.rb` where `test.rb` is
`Foo::Bar::Baz`

Before:

```
Pr(
  [ExprStmt(
     N(
       IdQualified(
         {name_last=(("Baz", ()), None); name_top=None;
          name_middle=Some(QExpr(
                             N(
                               IdQualified(
                                 {name_last=(("Bar", ()), None); name_top=None;
                                  name_middle=Some(QExpr(
                                                     N(
                                                       Id(("Foo", ()),
                                                         {id_info_id=1; id_hidden=false; id_resolved=Ref(
                                                          None); id_type=Ref(
                                                          None); id_svalue=Ref(
                                                          None); })), ()));
                                  name_info={id_info_id=2; id_hidden=false; id_resolved=Ref(
                                             None); id_type=Ref(None); id_svalue=Ref(
                                             None); };
                                  })), ()));
          name_info={id_info_id=3; id_hidden=false; id_resolved=Ref(None); id_type=Ref(None); id_svalue=Ref(None); }; })),
     ())])
```

After:

```
Pr(
  [ExprStmt(
     N(
       IdQualified(
         {name_last=(("Baz", ()), None); name_top=None;
          name_middle=Some(QDots([(("Foo", ()), None); (("Bar", ()), None)]));
          name_info={id_info_id=3; id_hidden=false; id_resolved=Ref(None); id_type=Ref(None); id_svalue=Ref(None); }; })),
     ())])
```