[PHP] Use latest semgrep-php fix with fix for echo
This will help https://github.com/returntocorp/semgrep/issues/3941

The parsing-stats for PHP in ocaml-tree-sitter-semgrep are now at
99.97%!

However, we are back at 92% for the parsing-stats in semgrep-core
for PHP (weird, because we were at 96% last week). Probably
some TODOs started to trigger more because of the parsing
fix.

test plan:
test file included
make test