feat(rust): Support method chain patterns (#5327)
* Update Rust tree-sitter parser

Pull in changes to the tree-sitter parser for syntax extensions, namely
ellipsis in method/field chains.

* Add translation for Rust method chain ellipsis

* Add tests for Rust ellipsis method chaining

* Update changelog

* Add tests for field chains

Currently we have a redundant match for the let expression line because
we translate a Rust let expression as `OtherExpr`, and when matching
against an `OtherExpr` we try all possible combinations. This seems like
the incorrect translation, but also probably out of scope for initially
adding this feature.