Handle PHP use imports without explict alias
For cases where an alias is not set, lower this to an `ImportFrom`
during the PHP to Generic AST translation layer. This is consistent with
how we handle the equivalent Java code, which follows a similar pattern
of language specific -> generic ast translation.

See, e.g., the second use in example #1 at
<https://www.php.net/manual/en/language.namespaces.importing.php>.