tainting: Analyze lambdas in context
Lambdas were already analyzed in isolation, as they are function
definitions, but they were not being analyzed in the context surrounding
them. So if e.g. some variable was tainted outside a lambda, and reached
a sink inside the lambda, this was not being detected.

Helps PA-1224
Closes PA-1308
Closes #5239

test plan:
make test # added two tests