Be more permissive of whitespace for xml ellipsis
In certain contexts we might have leading/trailing whitespace for an
ellipsis in a text or raw text parse node, but we probably ought to
treat this as an ellipsis regardless.

A better solution would probably be to generally migrate from having
ellipsis be an identifer to having it be a special variant.