deep: taint-mode: Always report taint findings on the sink match
Helps PA-1484
Follows: 1d5db279508 ("deep: taint-mode: Record call chain to reach sink (#4836)")

We experimented with matching always code in the function being analyzed,
but that seems to be more confusing for people. Semgrep users expect
matches on the sink, and then the call trace should explain them how the
taint is reaching the sink.

Besides, this also caused some "duplicate" findings that introduced
regressions in semgrep-rules, e.g. if every `request` is tainted:

    def f(request):
      # This is reported
      # ruleid: test
      sink(request)

    def g(request):
      # This is also reported !!!!! Regression ?????
      # `f` is called with `g`'s request which is also tainted, and that
      # taint reaches the `sink` too...
      f(request)

Whereas the user seems to expect matches only on the `sink`'s line.
The test suite reports a regression because it doesn't expect that extra
match.

(Note that this only concerns DeepSemgrep, in the good-old Semgrep this
issue doesn't apply because our FOSS taint tracking is intraprocedural!)

test plan:
make test