Parse booleans as constants
Closes https://github.com/returntocorp/semgrep/issues/5509

This allows constant propagation to happen on booleans in PHP by parsing
booleans as constants instead of Id. It adds a new construct to the AST for
bool. This was decided after reading the manual description of boolean, which
does not specify it as equivalent to an int value (see comment).

Test plan: make test