fix: scanning wrong thing in github actions (#5453)
GitHub Actions exposes a file with Event Webhook information and we relied on the sha of
the base to calculate the mergebase of the compare head and the base head. According to
GHA documentation, the event[pull_request][base][sha] is actually frozen at PR open time (though
investigations show it does get updated at least when you merge base into head but does not update
if any new PRs are merged onto base or if base is pulled into head through a third branch).

Because of this, any merges from base to compare branch will result in use not using the updated
merge-base. This results in semgrep reporting findings that were introduced in the base branch
after PR creation time.