Parse booleans as constants (#5521)
* Parse booleans as constants

Closes https://github.com/returntocorp/semgrep/issues/5509

This allows constant propagation to happen on booleans in PHP by parsing
booleans as constants instead of Id. It adds a new construct to the AST for
bool. This was decided after reading the manual description of boolean, which
does not specify it as equivalent to an int value (see comment).

Test plan: make test

* Changelog

* Fix pfff precommit

* Nat's suggestion

Co-authored-by: Nat Mote <nat@r2c.dev>

* Resolve a particularly annoying merge conflict

Co-authored-by: Nat Mote <nat@r2c.dev>