fix(js): Parse `New` as a separate construct from `Call` (#5510)
See the corresponding `pfff` commit for background.

Note that this changes the output on a `semgrep-rules` test. It looks
like this is because the `Call` pattern also happened to match `New`
expressions, since `New` expressions were represented as `Call`s. I'm
not sure how to address this. The new behavior seems correct, but it is
also clear that the rule author intended the rule to match both
constructor calls and function calls. I would appreciate reviewer
feedback on this issue.

Corresponding semgrep-rules PR:
https://github.com/returntocorp/semgrep-rules/pull/2130

Test plan: Automated tests