fix: Match TyN pattern against TyExpr (N ...) source (#5540)
* test(ts): Typed metavar for class test

* fix: Match `TyN` pattern against `TyExpr (N ...)` source

See the changed test for the goal. In JS, constructors can be arbitrary
expressions, so they are parsed into `TyExpr`. However, that leads to
missed matches.

The obvious solution is to desugar `TyExpr (N ...)` to `TyN` while
constructing the AST. I first tried to do this by modifiying
`AST_generic_helpers.expr_to_type` to do this desugaring, but I ran into
the regression test added in
https://github.com/returntocorp/semgrep/pull/4682.

It would be nice if there were a point where we could do this
desugaring, rather than having to handle it specifically in matching.
I'm not sure where it would make sense to do so, though.

Test plan: Automated tests