tainting: Yoann's review of returntocorp/semgrep#5547 (#5561)
Follows: 163ceee7656 ("tainting: Add experimental taint-propagators feature (#5547)")

Note that we tried to solve this with a hack in returntocorp/semgrep#5150
but it caused a bunch of FPs in semgrep-rules. The hack was essentially
to assume that in `x.f(y)` taint always propagated from `y` to `x`.

The typical FP was a call that incorrectly tainted an object or module,
that also happened to be part of a sink specification. For example, in
rule ruby.rails.security.audit.avoid-tainted-shell-call the `Shell` class
does not really get tainted even if we call `Shell.cat` on tainted data:

    # ruleid: avoid-tainted-shell-call
    Shell.cat(params[:filename])

But with the hack, `Shell` becomes tainted. Later on, when we call
`Shell.cat` on safe data, it triggered an FP. Why? Because the entire
`Shell.cat(...)` was marked as a sink, and `Shell` was considered
tainted!

    # ok: avoid-tainted-shell-call
    Shell.cat("/var/log/www/access.log")

Most of these FPs could be prevented by fine tuning pattern-sinks. But
anyhow it's clearly incorrect to taint `Shell`, so a better solution was
needed (hence `pattern-propagators`).

test plan:
make test # added one more test