Change rule type to use polymorphic variants (#5528)
* Change rule type to use polymorphic variants

Instead of using a normal variant for the mode field this changes the
rule type (`Rule.rule` -> `'mode Rule.rule_info`; `Rule.rule` is now an
alias for this with `'mode` as either a polymorphic variant for search
or taint) to be parameterised over a type variable `'mode` which is
instantiated to be specific polymorphic variants depending on the mode
of the rule.

This allows us to have functions which take/return rules where the modes
are restricted to a subset of all possible modes. Having this capability
is useful for if/when we have additional modes and, e.g., we want to
have a function take both search and taint rules but not rules having
other modes.

This just does the initial refactoring to change the main rule type and
some low-hanging fruit around the
`Match_{search,tainting}_mode.check_rule` types. There is probably some
additional refactoring which can be done to avoid redundant matches.

This PR also has a minor change to the CLI makefile to avoid a build
issue relating to recent directory renaming.