Stop using union-merge for CHANGELOG.md
Union-merges are helpful in some cases... but they also cause the
changelog to become incorrect with virtually every release PR that is
merged. Unfonrtunately union-merge is quite dumb. Perhaps it's better
to deal with the conflicts manually but do it well.

Reverts: ffba1572570 ("Merge changelog using union conflicts by default (#3379)")