fix(deep): Retry match without resolved name (#5582)
Instead of using empty ID info when retrying the match, just set the
name to be unresolved. This way, the other information, like the type
and svalue, is preserved. This allows the correct information to be
bound to the metavariable when the recursive call eventually calls
`envf`.

This fixes the equivalence_constant_propagation3.js test when run under
DeepSemgrep.

Test plan: Automated tests both here and with this change applied to
DeepSemgrep