Add `-pp` flag to core to pass extract rule file
It's unclear to me if this should really be a separate flag or not,
given that I'm implementing it using a mode. This introduces a bit of an
artificial barrier and will be a little bit less ergonomic when it comes
to making cli changes later, but is fine for testing and we can perhaps
change to sticking with just using multiple invocations of `-rules` later.