feat: Expose taint traces in semgrep-core text output (#5662)
This isn't plumbed through to the CLI yet, so it's not user-facing. But,
it's a step in that direction. The raw semgrep-core output is mostly
only used by Semgrep developers.

Instead of using the taint trace data separately provided to the match
hook (which isn't the codepath that JSON output uses), this change makes
it part of the `Pattern_match.t`. This involved defining a separate type
for taint traces that is more appropriate for match reporting, and
avoids a circular dependency. This will allow us to fairly easily add it
to the JSON output too, which is consumed by the CLI and the App.

Because this changes how the taint trace is accessed from the match
hook, a corresponding change to DeepSemgrep (which consumes taint trace
data there) will be required.

Test plan:

Run semgrep-core on the `tainting_rules/js/eslint_obj_inj.js` test:

```
eslint_obj_inj.js:4 with rule tainting
     return o[a]
  * Taint comes from:
    eslint_obj_inj.js:2
         var a = x
  * These intermediate values are tainted: a
  * This is how taint reaches the sink:
    eslint_obj_inj.js:4
         return o[a]
eslint_obj_inj.js:10 with rule tainting
     var z = o[b]
  * Taint comes from:
    eslint_obj_inj.js:8
         var b = baz(0)
  * These intermediate values are tainted: b
  * This is how taint reaches the sink:
    eslint_obj_inj.js:10
         var z = o[b]
eslint_obj_inj.js:21 with rule tainting
     return o[c]
  * Taint comes from:
    eslint_obj_inj.js:17
             c = x
  * These intermediate values are tainted: c
  * This is how taint reaches the sink:
    eslint_obj_inj.js:21
         return o[c]
```