Improve error message for mode-illegal properties
For properties which are allowed generally, but disallowed for a
specific mode, specifically indicate these as disallowed as invalid for
the current mode. There is room for future improvement in these error
messages in the case where a property is only used by one mode, in which
case it should just be a extra property on that specific mode, instead
of a top-level property.