feat: add `extract` as a new mode (#5621)
* Be more permissive of whitespace for xml ellipsis

In certain contexts we might have leading/trailing whitespace for an
ellipsis in a text or raw text parse node, but we probably ought to
treat this as an ellipsis regardless.

A better solution would probably be to generally migrate from having
ellipsis be an identifer to having it be a special variant.

* Derive show for many more types

Primarily this will help when printing values out to debug

* Add extracting module

Handles the logic for getting the extracted ranges from a source text
given an extract rule and determining how to map matches in the
extracted portion back to matches for the original text.

Currently just handles each extracted portion as a separate document.

* Implement logic for running extract rules

* Update cli progress for core-generated targets

* Document `Extract` module; add LGPL header.

* Disallow extract mode keys for other modes

* Don't require message and severity for extract

Currently just emits fake data for these fields. Futuristically these
could be moved to only the modes which require them.

* Handle no progess bar case in CLI

Updates the stdout processing to work even when a progress bar is not
produced (namely, when the plan contains only one target).

* Add severity fallback in CLI

This should be improved, but allows rules whose findings will not be
reported to omit severity and simply have a "dummy" internal value. This
will have some adverse interaction with severity filtering and should be
changed. Perhaps we should have another internal severity for rules like
these?

* Improve rule schema and update jsonschema

Updates jsonschema from 3.2 to 4.6.1 (types are 4.6.0).

Refactors our jsonschema from being `oneOf` many different mode
properties to instead satisfy `allOf` many implications on modes---this
has the effect of simply assuming the mode of the rule, which improves
diagnostics by avoiding dumping a list of irrelevant or illegal fields.

* Update pipfile for jsonschema transitive deps

* Update changelog

* Improve error message for mode-illegal properties

For properties which are allowed generally, but disallowed for a
specific mode, specifically indicate these as disallowed as invalid for
the current mode. There is room for future improvement in these error
messages in the case where a property is only used by one mode, in which
case it should just be a extra property on that specific mode, instead
of a top-level property.

* Improve error messages for mutually exclusive keys

* Improve handling for when no plan is used

In the case a plan is not used we do not calculate the number of
targets. This changes the handling around capturing the json output when
the number of targets is not calculated by Python.

* Update pytest snapshots

* Add some redundant properties to improve errors

* Allow minor version bumps for jsonschema

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>