test(ts): Add test for complex dots import pattern (#5695)
This illustrates the problem at the root of #5305. The issue is that we
desugar import statements that import multiple things into separate
import statements that each import a single thing. Combined with the
ellipses in the import list (which are just dropped, see
https://github.com/returntocorp/pfff/commit/045e544ee7fd83ee0dd7e71f57bcb02d1dfbd09e),
this leads to unexpected behavior.

We parse this pattern as a default import of `Foo`, followed by a named
import of `x`. Unfortunately, if there is an intervening import (`y`, in
this example), then the pattern doesn't match.