Reduce memory usage when not necessary
(See also the previous commits). This PR discards information that the user
won't see much earlier in the process; in most cases as soon as possible.
This is a change from previously, where we collected the information and
only decided at the end what to include in the result. That extra
information represents the majority of memory used by large repos.

In angular/angular with p/default, running with full verbosity (debug) results
in 3.34 GB used on exit, whereas running with only time information uses 1.67 GB
and running with no debug information uses 1.52 GB. (In future PRs, we expect to
be able to drop the 1.52 GB to below 1 GB)

Test plan: I believe `-json_time` is tested somewhere, but I need to check. I
don't think we have any tests for the correctness of skipped targets but I really
do not care very much about skipped targets tbh