infra: add marks to unmarked tests
Given the runtime of the overall test suite (less than when these were
introduced) and the fact that we have a durations file
<https://github.com/returntocorp/semgrep/blob/develop/cli/.test_durations>
we can probably just handle this programatically anyway.

However, because this currently blocks `make` at the top level on
develop, we should at least fix it as we currently have it configured.