Reduce memory usage when not necessary (#5730)
* Initial changes to the Report type

This splits out the result record types to save only the fields being used in
the CLI/CI output. It will make it easier to include or exclude fields for
memory use.

* String the changes through the rest of the code

* Reduce memory usage when not necessary

(See also the previous commits). This PR discards information that the user
won't see much earlier in the process; in most cases as soon as possible.
This is a change from previously, where we collected the information and
only decided at the end what to include in the result. That extra
information represents the majority of memory used by large repos.

In angular/angular with p/default, running with full verbosity (debug) results
in 3.34 GB used on exit, whereas running with only time information uses 1.67 GB
and running with no debug information uses 1.52 GB. (In future PRs, we expect to
be able to drop the 1.52 GB to below 1 GB)

Test plan: I believe `-json_time` is tested somewhere, but I need to check. I
don't think we have any tests for the correctness of skipped targets but I really
do not care very much about skipped targets tbh

* Remove mem tracing. (Should I add this as an option?)

* Modify extract to only compute skipped targets if helpful

* oops added Extract by mistake

* Fix Test_engine

* Update snapshot because we changed logging

* Use a record in Time to clarify that profiling is the same field as in Debug

* Allow parse errors

* Make skipped targets explicitly optional

* Update generated files

* Changelog

* Merge semgrep interfaces changes