fix: Match metavariables against resolved names
This undoes #5437, which turned out to be a misguided attempt to cut
down on duplicate matches. Perhaps we could ignore the metavariable
binding information when we dedup, though that may come with its own
problems. For now, we'd rather deal with duplicates than fail to match
in cases where we should.

Test plan: Automated tests

Fixes #5690