feat(cli): Display dataflow traces in text output (#5751)
This introduces a `--dataflow-traces` flag to the `scan` command. The
default behavior is unchanged, but when this flag is passed and the
output mode is text, the dataflow trace is also displayed. For now, it
is only available for taint mode findings, and includes the source of
the taint as well as intermediate variables through which the taint
flows to the sink.

The JSON output already includes this information.

In support of this change, I extracted the `get_lines` function into
`util.py`, and the `_format_lines` function out of `_finding_to_line`.

There are a bunch of edge cases and options when printing matches. I
wasn't sure how to exercise all of them, and I'm not at all confident
that this will look right for every possible input. However, I do
believe that the default behavior is unchanged, and since this is behind
a flag it's less important to get all of the edge cases right. We can
refine it as we get feedback from users who are motivated to use the
flag, and perhaps eventually remove the flag and make dataflow trace
display the default.

We already plan to display this data by default in PR/MR comments and
perhaps in the playground, so it would make sense to eventually display
it on the CLI as well. However, several people expressed concern that it
could get too verbose, so it is behind a flag for the moment.

Test plan: Automated tests for ensuring default behavior remains as-is

Manual test of `--dataflow-traces` flag:

```
(cli) $ python -m semgrep scan -c ../semgrep-core/tests/OTHER/rules/taint_lambda.yaml ../semgrep-core/tests/OTHER/rules/taint_lambda.js
Scanning 1 file.

Findings:

  ../semgrep-core/tests/OTHER/rules/taint_lambda.js
     semgrep-core.tests.OTHER.rules.test
        Test

          8┆ console.log(numbers.map(i => sink(i + x)));

Some files were skipped or only partially analyzed.
  Scan skipped: 1 files matching .semgrepignore patterns
  For a full list of skipped files, run semgrep with the --verbose flag.

Ran 1 rule on 1 file: 1 finding.
(cli) $ python -m semgrep scan -c ../semgrep-core/tests/OTHER/rules/taint_lambda.yaml ../semgrep-core/tests/OTHER/rules/taint_lambda.js --dataflow-traces
Scanning 1 file.

Findings:

  ../semgrep-core/tests/OTHER/rules/taint_lambda.js
     semgrep-core.tests.OTHER.rules.test
        Test

          8┆ console.log(numbers.map(i => sink(i + x)));
        Taint comes from:
          6┆ var x = source;
        Taint flows through these intermediate variables:
          6┆ var x = source;

Some files were skipped or only partially analyzed.
  Scan skipped: 1 files matching .semgrepignore patterns
  For a full list of skipped files, run semgrep with the --verbose flag.

Ran 1 rule on 1 file: 1 finding.
```