tainting: Field sensitive taint-mode
Closes PA-1278

Make taint analysis sensitive to individual fields in records/objects in
a limited way. Essentially, it adds _per variable_ field sensitivity and
not _per object in memory_ field sensitivity. There is no alias analysis
involved. A summary of new behavior

 1. Taintedness/cleaness is tracked individually for each field of each
    variable.
 2. If a field `a.b.c` is marked tainted, then `a`, `a.b` and `a.b.c` are
    all marked tainted.
 3. If a field `a.b.c` is marked clean, then `a.b.c` and any previously
    tainted field of which `a.b.c` is a _prefix_ is marked clean. So
    `a.b.c.d` will now be clean, even if had been previously tainted.
 4. When checking the taintedness of some dot expression `a.b.c.d`, we
    use the most up to date knowledge of its fields. That is, if we know
    only that `a.b` is tainted, then `a.b.c.d` is tainted. But if we know
    that `a.b.c` was marked clean after `a.b` was tainted, then `a.b.c.d`
    is clean.

The following is technical changes are needed to enable this behavior:

1. IL lvals now have their offsets represented as _backwards lists_.
   So `a.b.c` is `{base = a, offset = [.c ; .b]}`. This allows even
   nested field access lvals to be represented as single variables.
2. We differentiate between a variable with _unknown taint_ and one
   with _known lack of taint_. To see why this is needed, consider:
    ```
    a = source
    a.b = safe
    sink(a.b)
    ```
    If `a.b = safe` simply removed the string corresponding to `a.b`
    from the set of tainted variables, it would do nothing. We need
    to specifically remember the fact that `a.b` is clean!

One major short-coming, other than the lack of alias analysis, is that
when a variable is marked tainted because one of it's fields was tainted,
we do _not_ remember what fields caused the taint. So if these fields are
later marked clean, it does not erase the taint of the top level variable.
For example:
```
x = safe // x not tainted
x.y.z = source // x is tainted
x.y.z = safe // x still tainted!
```

test plan:
make test # added 4 tests

Edited-by: Iago Abal <iago@r2c.dev>